timer variable controlled
chaning the timer to use variables to configure timer of(minutes portion does not seem to work)

+++ Inc/Timer.h
@@ -6,4 +6,4 @@
  */
 #define INT_16BIT_MAX 65536
 
-int GetDesiredPeriod(int DesiredDelay,int currentPrescaler);
+int GetDesiredPeriod(float DesiredDelay,int currentPrescaler);
